lupdate: Support for Typed Enum in CPP Parser
The CPP parser of lupdate gets misleaded when visiting a COLON token
after an enum declaration. This is the case when a cpp code contains
typed enum like the following example:

`enum class MyEnum: int { ... }`

In this case when the parser visits COLON, it assumes it's parsing
a class constructor initializer.

Allow the parser to flexibly detect and ignore colons in front
of the typed enums.

Fixes: QTBUG-128904
Change-Id: I5331c367de2b93d7097eca5418d3835344c3362a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>