qdoc: Use full paths for class/namespace reference page titles
Instead of documenting the full scope of a class in a subtitle of a
reference page, list it already in the page title.

In HTMLGenerator, continue producing also a subtitle if there are
either documented parent(s) to link to, or if the reference is for
a templated class. In this case, the template declaration is visible
in the subtitle.

Use full paths in All/Obsolete Members page titles as well.

In DocbookGenerator, there are no links within a subtitle, so generate
a separate subtitle only when there is a template declaration.

Add NamespaceNode handling to Aggregate::typeWord() to avoid hardcoded
"Namespace" string literals in title-generating code.

Fixes: QTBUG-129709
Change-Id: I218ddf1431bee875fb8be629af7f4588de55e841
Reviewed-by: <nicolas.fella@kde.org>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>