lupdate CPP Parser handling templates
Whenever the CPP parser of lupdates observes a class_token
eventually followed by a namespace_token
("::"), it assumes a new context (namespaced class) is starting.
Hence, the parser tries to qualify the identifiers that are
tied together by "::". This is usually correct and desired behavior.
For instance:

namespace MyNameSpace {
	class MyClass;
}

class MyNameSpace::MyClass {
...  // context MyNameSpace::MyClass
}

But this fails if we have a class_token inside a template entity
followed by the namespace_token:

template <class = typename std::string>

or

template <class T, class = std::enable_if_t ...>

In these cases, neither of "std::string" and "std::enable_if_t"
can be qualified by the parser. As a result the user gets a
warning about the "definition of undeclared qualified class".

Allow the parser to identify the class keywords that are
in a template entity, ignore qualifying them, and only start
qualifying when the real class definition starts.

Fixes: QTBUG-130147
Change-Id: I2980db9cb4996885727680a9e92e2b948d2364ba
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>