lupdate: Warn the user on using template strings for translation
Warn about using template strings in translator functions:

qsTr(`The runtime value ${val} is fetched`)

When running the application, for val=100, the qsTr method will see:

"The runtime value 100 is fetched"

And this is different from what Linguist will see and what will be
produced in the .ts files. Hence, the translation cannot be found later.

Warn user about this situation and document the reasoning behind this
limitation.

Fixes: QTBUG-127527
Change-Id: I550fb76728be49c3d9778c260aca08aaaf7ea62f
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>