qdoc: QML Visitor: Handle namespaced imports
When visiting a .qml file for documentation comments, QDoc did not store
the `import Id` that the QML type uses for importing a module under a
specific namespace. This resulted in the possibility of resolving an
incorrect base type for the QML type, if multiple identically-named
types in different modules were visible to QDoc.

Store the import Id as provided by QQmlJS::AST into an ImportRec
structure, and use it in QDocDatase when searching for QML types under
the recorded imports.

[ChangeLog][QDoc] QDoc now resolves the correct QML base type imported
under a namespace ID.

Fixes: QTBUG-130734
Pick-to: 6.8
Change-Id: I219172b97ec17f30d6d66828abc14139d7cda2a4
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>