qdoc: Allow \relates for class/namespace members
The original purpose for the \relates command was to associate the
documentation of entities in the global scope to a page-generating
aggregate (e.g. a class or a namespace) reference.

This commit removes the global scope restriction for such entities,
making it possible to list, for example, a function in a documented
namespace also under the reference page of a class the function is
related to, under its `Related Non-members` section.

Make Aggregate::resolveRelates() run recursively, instead of
processing only the members of the global namespace Aggregate.

Remove the `<function> must be global` warning from \relates command
processing.

Adjust and improve documentation for the \relates command.

Fixes: QTBUG-130674
Pick-to: 6.8
Done-with: Nicolas Fella <nicolas.fella@kde.org>
Change-Id: I56913637e7d931dee6b0dcc90aa3abda3f697611
Reviewed-by: Paul Wicking <paul.wicking@qt.io>