QDoc: Stop passing $CLANG_RESOURCE_DIR to Clang
QDoc's `ClangCodeParser` passes a set of default arguments to Clang
when interfacing with it. These arguments include the C++ standard,
`-D` options, etc. Among these is
`-I/path/expanded/from/$CLANG_RESOURCE_DIR`. This path is generated by
CMake at build time and may therefore be incorrect if the binary is
moved to a different system; reports indicate that this indeed is an
issue in Yocto builds.

[ChangeLog][QDoc] QDoc no longer passes the contents of the
`CLANG_RESOURCE_DIR` environment variable as an include path when
interfacing with Clang.

Pick-to: 6.8
Task-number: QTBUG-130282
Change-Id: I3642a7e67f8de175491e3d7626821a66eb8dd86d
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>