qtattributionscanner: Extend LicenseId-based file search
Before this patch, the scanner LicenseId-based automatic search
only looked for <Id>.txt files in LICENSES directory.

If a license is only listed in a qt_attribution.json but not associated
to source code its presence in LICENSES directory causes reuse to error
out. Such a license file should be moved and renamed with a LICENSE.
prefix to avoid license check from reuse. Before this patch, the only
way for the scanner to find such a license was through the use of
LicenseFile(s) field in the qt_attribution.json.

LicenseFile(s) takes precedence over LicenseId, and if a license is
listed in LicenseId but not in LicenseFile(s) it will simply be omitted
without causing any error. To avoid omission, only LicenseId entry
should be used.

Extending the scanner LicenseId-based automatic search to LICENSE.Id.txt
file next to the qt_attribution.json allows avoiding conflict with reuse
without using LicenseFile(s) field.

Pick-to: 6.8
Change-Id: I0c46c02a1d899cd8a1ba75782c6da0c33ade995a
Reviewed-by: Kai Köhne <kai.koehne@qt.io>