qdoc: Refactor \qmlproperty argument parsing
QmlDocVisitor and CppCodeParser included the same logic (with slight
variation) for parsing the arguments passed to the topic commands for
documenting QML properties.

Rewrite and extract the logic to a separate container + static helper
function, and make it support both use cases; CppCodeParser expects
a property type and a name qualified with at least the parent QML type
name, whereas QmlDocVisitor accepts unqualified property names as the
QML type is known already.

If the parsing fails on incorrect input, make the helper function
return std::nullopt, instead of partially-populated fields as was done
previously. This exposed an issue in
CppCodeParser::processQmlProperties() where we potentially created
QmlTypeNode instances with incorrect/empty name. This is now fixed.

Pick-to: 6.9 6.8
Change-Id: I3353bc56b03504d0cfbdbf97595e34d594c1733c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>