Allow Empty Strings in the Translation Tools
Allow the user to use empty message texts in translation commands.
This can be useful if a user uses empty strings
for things that should show up only in some languages. For instance
a warning that only the English translation is an official one.

Provide support for empty source texts for the parsers in
python, java, js, qml, and cpp.

Also test the functionality for the following translation tools:
- lupdate (with and without clang parser)
- lconvert (round trips: ts-qm-ts and ts-xlf-ts)
- lrelease
- the linguist application.

Fixes: QTBUG-110936
Change-Id: I836c4a888d7495f602926a9b1ef2e1ffe524de29
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>