qdoc: Allow overriding QML property types with \qmlproperty arguments
Previously QDoc allowed to override the data type of aliased QML
properties only: The target property may be an undocumented
implementation detail, and therefore documenting the underlying type
directly was necessary.

However, there are other circumstances where the desired data type
as it appears in the documentation differs from the implementation
in the .qml file. As documenting QML properties outside .qml files
never had this restriction, remove it also when parsing .qml for
documentation comments to be consistent.

Note that the QML visitor stores the property list modifier to the
QmlPropertyNode instance already before parsing \qmlproperty
arguments; this modifier is then re-applied when overriding the
data type. In practice, this means that it's not possible to
override a list-type property to be documented as a non-list type,
only the underlying data type can be changed.

Pick-to: 6.9 6.8
Task-number: QTBUG-128327
Change-Id: I9282aef2c627b66d79f5b74f56e09fa77deab250
Reviewed-by: Paul Wicking <paul.wicking@qt.io>