qdoc: Stop overwriting docs for clashing \qmltypes in different modules
When parsing a \qmltype command, QDoc checks whether a QmlTypeNode with
the provided name and module identifier already exists. One of the
special cases is a placeholder node instance created by a previous
\qmlproperty command, which specified a QML type that had not yet been
created.

The checks performed before tying the \qmltype documentation to an
existing node were insufficient, however. When a node with identical
name was found in the primary tree, QDoc always assumed it was a
placeholder instance. However, the match can also be for an
identically-named type in a different QML module.

To fix, require the potential placeholder node to have an empty
module identifier before matching it with the documentation.

Pick-to: 6.8
Fixes: QTBUG-131572
Change-Id: I873d55247d9d531e187099545d524f83831d5e33
Reviewed-by: Paul Wicking <paul.wicking@qt.io>