lconvert: Resolve Translation Dependencies
lconvert ignores the dependency attribute of the translation
files when it's used to merge multple files into one. Currently,
the resulting translation file contains the dependencies
of the first provided file, plus the merged messages of
all the files.

The desired behavior is that if some of the files are
the dependencies of each other, they are satisfied in the
resulting file.

Resolve translation dependencies in lconvert by gathering all
dependencies from the provided files, and resolving them when
they are satisfied. This should be then reflected in the
resulting file.

Fixes: QTBUG-125951
Change-Id: I5a8e93130894d6f543e35a87b134e0176c5b33a9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Kai Köhne <kai.koehne@qt.io>