qdoc: Qt help project: Stop generating 'All members' for wrapper classes
Qt OpenGL includes a number of classes marked with the \wrapper QDoc
command; they do not document any members as they act as a wrapper
interface for functions from the OpenGL specification.

While the HTMLGenerator correctly skips the generation of 'All members'
file/link on the reference pages, the HelpProjectWriter generated a
section with a link to a non-existent file.

Ultimately, the decision whether to generate an 'All members' file is
done via the Sections class after distributing all members into
sections. Repeating this process in the qhp would be too expensive -
therefore, make a naive assumption that wrapped classes have no
associated 'All Members' file when generating a .qhp.

Pick-to: 6.8
Fixes: QTBUG-131375
Change-Id: I2735a7710bf3b57ce7db6ba22ee1061fa51bc1b8
Reviewed-by: Paul Wicking <paul.wicking@qt.io>