lupdate & linguist: Add support for NCR Notations
If the translation text in source code contains
"numeric character reference" (NCR) representation of
a character (e.g., "&#x200b"), allow lupdate to
write those characters in .ts files as they are written
in the source code without interpreting them.

On the Linguist side, to support this feature, add an
option to Linguist view menu to toggle displaying
these characters in the form of "numeric
character reference", again without interpreting them.
When the option is enabled, Linguist should only change
the displayed text (no effect on the
output .ts file on Linguist).

Fixes: QTBUG-130609
Change-Id: Icb14a933bbe4743a9aea0e41c35c7dac53156ad0
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>