lupdate: Escape Unicode Chars for Different Parsers
The unicode characters are escaped correctly in the cpp parser.
But for the Python parser the escaped chars
were not handled correctly by parser.

Add the support for handling escaped unicode chars for Python parser.
Parsers of Java, JS, and QML, do not need to take \U escaping into
account since there is no such a concept as \U escaping in those
languages.

Change-Id: I3164f4a23d50033d1b0b2aa0bfa6a7897cc6a4ba
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>