Linguist: Disable Warnings for NCR Characters
Linguist shows warnings when the source text contains "&" but the
translation doesn't contain it, or vice versa. This is mostly
the desired behavior, except if "&" follows to define a NCR
character: "&x200b". In this case, because of different
characteristics of different languages, in some cases, it's ok to
have NCR characters in translations and not have it in the source,
and vice versa.

Remove the warning from Linguist when NCR characters are defined.

Fixes: QTBUG-131951
Change-Id: I466be5a9ee7e910bf8bac83e7a98af91c5579e2b
Reviewed-by: Kai Köhne <kai.koehne@qt.io>