qdoc: Allow documenting QML property groups in .qml
QmlDocVisitor::applyDocumentation() was capable of processing
multiple topic commands in the same QDoc comment block when
parsing documentation in a .qml file. It did not understand
QML property groups though, always treating each \qmlproperty
topic individually and duplicating the doc content for each
in the generated output.

After collecting the resulting Nodes from topic commands,
check whether they qualify for a QML property group
documentation: There must be at least two properties to
form a group, and each must be named with the correct
group prefix, delimited by '.'. The name of the group is
the actual declared QML property name in .qml.

If we have a property group, construct a SharedCommentNode
based on the node list. If not, process each Node separately
as was done previously.

Consolidate Node::setLocation() and Node::setDoc() calls into
this new 'post-processing' step. As a side-effect, we now get
correct location info passed to each Node instance and
stored into the .index file as well.

Fixes: QTBUG-132151
Pick-to: 6.9
Change-Id: I0c6e108bb2c7fe76ac1091ca733428593a1c7060
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: <nicolas.fella@kde.org>