QDoc: Write title attribute for images in HTML
When generating HTML output, QDoc generates an alternate text for
images if one is provided. QDoc also offers a way to add an image
caption, by way of the `\caption` command. However, there is no way to
specify a text that is to be used as a tooltip/hover text for images.

The usefulness of this is debatable, and it's known to possibly have
negative impact on the user experience for users that use accessibility
products such as screen readers. It's also if little to no value for
users that navigate the documentation with touch screens (e.g. tablets
or mobile phones), keyboard, or any other means beside a mouse.

Researching the implementation has shown that there's no consensus on
what is the "correct" way to provide tooltips for images, in that the
use of the `title` attribute for images is contended. Some screen
readers read the contents of both the `alt` attribute and the `title`
attribute, which may cause confusion for users if the content is
duplicated. Some screen readers /do not/ read the contents of the
`title` attribute, such that it cannot be relied upon to convey
information of any importance. Thus; while some argue that the contents
of the `alt` and `title` attributes must match, while others argue that
they must /not/ match. Finally, some think the `title` attribute is
nonsensical altogether.

This change is an attempt at providing the feature as requested by users
without imposing unwanted behavior on those that do not want it, while
minimizing the scope of the change:

- Add a new configuration variable, `usealttextastitle`.
- Add an early return for empty strings in `protect()`. This allows for
  removing the now redundant TODO comment. It also ensures an early
  return if applicable, as callers do not consistently perform the
  check, and allows for flattening the branch.
- Modify the code that generates the image embed element such that it
  includes a `title` attribute, provided that the new configuration flag
  is set.
- Extend the output test for QDoc with a trivial test project that makes
  use of the new configuration variable and feature.
- Update the QDoc manual with an entry for the new configuration
  variable, and modify the section "Including Images" to reflect a more
  modern approach to accessibility.

[ChangeLog][QDoc] QDoc can now generate a `title` attribute for your images.
Set the configuration variable `usealttextastitle = true` in your .qdocconf
file.

Pick-to: 6.9
Fixes: QTBUG-132432
Change-Id: Iee175ed325753702693f3b52e0aeb09f0cd9d9f0
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>