CMake: Add option to skip extraction of languages from ts files
The behavior of extracting languages from ts files was mostly needed
for user projects that ship macos / ios bundles, so we know what value
to assign to the bundle CFBundleLocalizations key.

For the case of processing the qttranslations repo, this is not
needed.

Add an opt out that can be set at directory-scope.
It will be set in qttranslations CMakeLists.txt.

Pick-to: 6.8 6.9
Change-Id: I47e29d5f119ca621cfe3873dc9cabf17aac3d88d
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>