linguist: prioritize focused editors when using CTRL+A
when using CTRL+A to select text from a widget, prioritize an
editor that has the focus, over a mouse-hovered but non-focused,
field.
The current behavior prioritizes the hovered source text widget,
over a focused editor, which is confusing for users.
With the new change, the hovered source text widget could still
be selected, but only if the focus is not on an editor.

Fixes: QTBUG-57734
Change-Id: Idcbd8e87982ce0303d5248349aa49ab3a2e25c96
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>