Help: qthelpheadless feature for QHelpEnginePlugin
Introduce a qthelpheadless feature to strip-down Help of Qt::Gui and
Qt::Widgets stuff. This should allow QHelpEnginePlugin to be built when
Qt is configured with -no-gui.

This is needed so that qmlls can statically link the help plugin without
exploding its binary size by linking against the Gui and Widgets stuff
that is not needed by qmlls.

This allows to shrink the qmlls binary from 17MB to 13MB.

Task-number: QTBUG-133106
Change-Id: I756c2cfa336e661af9715977c883c1e345a58398
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>