linguist: apply validation on different translation variants
linguist only validates the translations of the first length
variant. As a result, when a user has multiple variants and toggles
different validations in linguist, they only work on the first
variant and ignore the rest of the translations.
Perform validation on all translation variants based on
the user preferences.

Change-Id: I93ce0e1131e5dfed6f1b3f9cb1bc65b0ab9baa78
Fixes: QTBUG-52661
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>