linguist: protect message ID characters in XLIFF files
If an XLF file contains message ids with special characters
(e.g., &quot;), by saving the file in linguist, the characters are
replaced by their evaluated representations. This makes the file
invalid for opening in linguist. As a result, even opening and saving
such an XLF file in linguist without changing anything,
makes it invalid.
linguist calls protect(...) to adjust the characters of message ids,
and texts for other formats. Only in XLF for message ID it seems to
be missing.
Add the protect(...) call for message ID when writing an XLF file in
linguist.

Fixes: QTBUG-68098
Change-Id: Ib2f2375cfb9e85e38688c190d72a991309efc32a
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>