lupdate, lconvert: provide an option to sort messages in a context
lupdate and lconvert keep the order of messages as their appear in the
code. This is usually the desired behavior. Except, when a user wants to
compare two versions of a TS file (caused by messages rearranged to
different line numbers). Then, the comparison produces false positives.

Allow the user to sort messages in a context alphabetically, in lupdate
and lconvert, by providing a command line option "-sort-messages".

Fixes: QTBUG-36443
Change-Id: I0325c1d27d49243dc284b129895456ffa66aa962
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>