qdoc: Fix auto-tying private signal documentation in headers
When QDoc's ClangCodeParser visits a header, it omits a signal parameter
of type `QPrivateSignal` as this argument is not to be visible in the
generated documentation. Instead, the function node is marked as a
private signal which appears in its meta-information.

This works as expected when documenting a private signal with an \fn
command, assuming the argument is also omitted in the \fn signature.
But when documenting the signal directly above its declaration in the
header without an \fn (auto-tie), QDoc ended up comparing the number
of arguments between the declaration in Clang AST (where the argument
exists), and the stored FunctionNode (where it was removed), resulting
in a match failure.

Fix this by ignoring the last argument when matching a private signal
in cases where the argument was dropped.

Fixes: QTBUG-132881
Pick-to: 6.9
Change-Id: I9199b7f69f01324e747c287d047f56b310b3a60d
Reviewed-by: Paul Wicking <paul.wicking@qt.io>