linguist: make statistics menu item non-checkable
The View->Statistics menu item does not need to be checkable
since its sole function is to open a dialog displaying statistics.

Currently, the menu item appears checkable. When selected, it
opens the statistics dialog, and upon closing the dialog, the
item is automatically unchecked. This behavior is misleading,
as it suggests a persistent state rather than a one-time action.

Fixes: QTBUG-10581
Change-Id: Ia6b4d2b5c78aee4e31ba554301f74de899153858
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>